Protects against MediaMTX 1.12.0 syntax check
Addresses an issue with MediaMTX 1.12.0's aggressive syntax checking
for `recordPath`, which applies validations the setting even when recording
is disabled.

To prevent errors, the code now checks if the computed `record_path` contains *either* an`"%s"` or **all** of `"%Y"`, `"%m"`, `"%d"`, `"%H"`, `"%M"`, `"%S"` format specifiers. If not compliant, it appends `"_%s"` and emits a warning.

Additionally, the default `RECORD_PATH` and `RECORD_FILE_NAME` are updated.