Add exception handling to ffmpeg pruning logic
This will prevent exceptions in the pruner (which could happen if two snapshots are requested in an overlap of the prune cycle... so one tries to delete a file another has already removed). These exceptions were causing the entire snapshot thread to die.

Also, made a change to keep a set of the file parents so we can only test if they need to be deleted (if empty) only AFTER pruning the entire set of files. This will ensure that we cleanup quicker.